Factor a barrel of monkeys out into a template
There are many names for the grouping of monkeys, such as:
* Troop
* Carload
* Tribe
* Barrel

When choosing between the above grouping names, the names "tribe" or
"troop" invoke a large number of monkeys, which seems inappropriate when
using two replicas of a monkey in a Kubernetes StatefulSet. Likewise, a
carload of monkeys would typically number more than two; an average sedan
would fit at least four monkeys quite comfortably, or perhaps more
depending on the size of the monkeys.

The final option, a "barrel", seems more appropriate. While the typical
barrel can hold between 100-160L of goods (though the standard varies
depending on the specfic good), the actual size of the barrel is often
more in the range of a little less than one metre tall by about half a
metre wide. Without knowing the particular species of monkey, it is
difficult to make an estimation regarding the volume constraints for
fitting monkeys in such a barrel. If we were to measure monkeys by
volume, then it certainly seems plausible to fit more than two monkeys
in a barrel, particularly if they are Pygmy monkeys. That said, there
are other constraints to be considered as well. The male mandrill for
instance may be as tall as 1 metre without including the tail, so even
if it is possible to fit such a monkey into a barrel, it may not be
sufficient to ensure the well-being of said monkey. Therefore, it
appears to be an appropriate usage of the grouping of a "barrel" of
monkeys to represent a StatefulSet with two replica monkeys, for usage
by chaos tests which require a pair of monkeys to work in tandem.

This generic template YAML is used initially by the Clique chaos test,
and will be used by another upcoming chaos monkey which requires a pair
of monkeys to operate together.

Signed-off-by: Joe Stringer <joe@covalent.io>