add tennis-match monkey
Continuously create and delete services in one pod. In another pod, consume
these events from Kubernetes. If the events from Kubernetes cannot be consumed
due to the connection getting closed, the pod crashes. This is intended to test
the durability of Kubernetes Services plumbed into Cilium over a long period of
time.

To do so, add the following to the base chaos-monkey infrastructure:

* RBAC rules which allow full control over the lifecycle of Kubernetes Endpoints
and Services.
* Update `barrel-statefulset.yaml` to export the pod's namespace as an
environment variable to any monkey which uses the file as a template.

Signed-off by: Ian Vernon <ian@cilium.io>