lib: wrap monitor command into a shell to get correct pid
Getting the PID of the previous process via `$!` will only get the last process
that was started in a series of piped commands, e.g.:

```
$ p1 | p2 &
$ PID=$!
```

Will only get the PID of p2, not the whole command. Fix this issue in the code
that starts the monitor by wrapping the monitor command which is piped into `ts`
in a shell, whose PID is then obtained.

Signed-off by: Ian Vernon <ian@cilium.io>