add apple-tree monkey
This monkey likes CIDR's (get it? cider... it comes from apples).

This monkey has two pods: a controller which adds and deletes CIDR policies
which refer to the same CIDR, and another pod which the CIDR policy selects,
which queries the CIDR allowed by the policy. The monkey will complain via Slack
if it cannot connect to the external CIDR.

Signed-off by: Ian Vernon <ian@cilium.io>