FQDN policy with S3 files.
This is a new test that validates a FQDN policy with Amazon S3 URL that
updates the IPs frequently.

It will make a head request to the given target, if it replies with a
non 200 stauts code will fail. If in the given timeout cannot connect it
will fail.

For debugging porpouses the DNS lookup time will be printed also the
remoteIP,to make sure that is added in the identity list.

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>